+++ CHANGES.rst
@@ -33,6 +33,8 @@ Unreleased
 -   Use modern packaging metadata with ``pyproject.toml`` instead of ``setup.cfg``.
     :pr:`4947`
 -   Ensure subdomains are applied with nested blueprints. :issue:`4834`
+-   ``config.from_file`` can use ``text=False`` to indicate that the parser wants a
+    binary file instead. :issue:`4989`
 
 
 Version 2.2.3